+++ src/features/offer/components/MovieScreeningCalendar/MovieScreeningCalendar.tsx
@@ -2,14 +2,10 @@ import React, { FunctionComponent, useEffect, useRef } from 'react'
 import { ScrollView } from 'react-native'
 
 import { OfferResponse } from 'api/gen'
-import { useAuthContext } from 'features/auth/context/AuthContext'
-import { useOngoingOrEndedBooking } from 'features/bookings/api'
 import { MovieCalendar } from 'features/offer/components/MovieCalendar/MovieCalendar'
 import { useMovieScreeningCalendar } from 'features/offer/components/MovieScreeningCalendar/useMovieScreeningCalendar'
+import { useSelectedDateScreening } from 'features/offer/components/MovieScreeningCalendar/useSelectedDateScreenings'
 import { useOfferCTAButton } from 'features/offer/components/OfferCTAButton/useOfferCTAButton'
-import { getBookingOfferId } from 'features/offer/helpers/getBookingOfferId/getBookingOfferId'
-import { getIsBookedOffer } from 'features/offer/helpers/useCtaWordingAndAction/useCtaWordingAndAction'
-import { useCreditForOffer } from 'features/offer/helpers/useHasEnoughCredit/useHasEnoughCredit'
 import { Subcategory } from 'libs/subcategories/types'
 import { EventCardList } from 'ui/components/eventCard/EventCardList'
 import { Spacer } from 'ui/theme'
@@ -22,13 +18,13 @@ export const MovieScreeningCalendar: FunctionComponent<Props> = ({ offer, subcat
   const { stocks, id: offerId } = offer
   const offerVenueId = offer.venue.id
 
-  const {
-    movieScreeningDates,
-    selectedDate,
-    setSelectedDate,
-    getSelectedDateScreenings,
-    bookingData,
-  } = useMovieScreeningCalendar(stocks)
+  const { movieScreeningDates, selectedDate, setSelectedDate, selectedScreeningStock } =
+    useMovieScreeningCalendar(stocks)
+
+  const { bookingData, selectedDateScreenings } = useSelectedDateScreening(
+    selectedScreeningStock,
+    offerId
+  )
 
   const { onPress: onPressOfferCTA, CTAOfferModal } = useOfferCTAButton(
     offer,
@@ -38,21 +34,7 @@ export const MovieScreeningCalendar: FunctionComponent<Props> = ({ offer, subcat
 
   const scrollViewRef = useRef<ScrollView | null>(null)
 
-  const { user, isLoggedIn } = useAuthContext()
-  const userCredit = useCreditForOffer(offerId)
-  const hasBookedOffer = offerId && user ? getIsBookedOffer(offerId, user.bookedOffers) : false
-  const { data: userBooking } = useOngoingOrEndedBooking(
-    getBookingOfferId(offerId, user?.bookedOffers) ?? 0
-  )
-
-  const eventCardData = getSelectedDateScreenings(
-    offerVenueId,
-    hasBookedOffer,
-    userCredit,
-    isLoggedIn,
-    onPressOfferCTA,
-    userBooking
-  )
+  const eventCardData = selectedDateScreenings(offerVenueId, onPressOfferCTA)
 
   useEffect(() => {
     if (scrollViewRef?.current) {
