+++ src/features/offer/components/OfferCTAButton/useOfferCTAButton.tsx
@@ -2,7 +2,7 @@ import { useRoute } from '@react-navigation/native'
 
 import { OfferResponse } from 'api/gen'
 import { UseRouteType, StepperOrigin } from 'features/navigation/RootNavigator/types'
-import { MovieScreeningBookingData } from 'features/offer/components/MovieScreeningCalendar/type'
+import { MovieScreeningBookingData } from 'features/offer/components/MovieScreeningCalendar/types'
 import { useCtaWordingAndAction } from 'features/offer/helpers/useCtaWordingAndAction/useCtaWordingAndAction'
 import { Subcategory } from 'libs/subcategories/types'
 import { useBookOfferModal } from 'shared/offer/helpers/useBookOfferModal'
