+++ src/shared/offer/helpers/useBookOfferModal.tsx
@@ -2,7 +2,7 @@ import React, { ReactElement } from 'react'
 
 import { BookingOfferModal } from 'features/bookOffer/pages/BookingOfferModal'
 import { StepperOrigin } from 'features/navigation/RootNavigator/types'
-import { MovieScreeningBookingData } from 'features/offer/components/MovieScreeningCalendar/type'
+import { MovieScreeningBookingData } from 'features/offer/components/MovieScreeningCalendar/types'
 import { ApplicationProcessingModal } from 'shared/offer/components/ApplicationProcessingModal/ApplicationProcessingModal'
 import { AuthenticationModal } from 'shared/offer/components/AuthenticationModal/AuthenticationModal'
 import { ErrorApplicationModal } from 'shared/offer/components/ErrorApplicationModal/ErrorApplicationModal'
